Rename useIsFirefoxOrSafari.ts to useIsFirefox.ts
