+++ hooks/useIsFirefox.ts
